Adding provider replication test and relocating a fixture
